requirements.txt
tensorflow
Keras-Preprocessing
scikit-image
matplotlib
Pillow
numpy
scikit-learn